A few tidy-ups
