--- aurora4spark-parent/aurora4spark-sql-plugin/src/main/scala/com/nec/spark/cgescape/UnsafeBatchProcessor.scala
@@ -1,9 +1,0 @@
-package com.nec.spark.cgescape
-
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.UnsafeRow
-
-trait UnsafeBatchProcessor {
-  def insertRow(unsafeRow: UnsafeRow): Unit
-  def execute(): Iterator[InternalRow]
-}
