+++ aurora4spark-parent/aurora4spark-sql-plugin/src/main/scala/com/nec/spark/cgescape/UnsafeExternalProcessorBase.scala
@@ -1,8 +1,10 @@
 package com.nec.spark.cgescape
 
+import com.nec.spark.cgescape.IdentityCodegenBatchPlan.UnsafeExternalDuplicator
+import com.nec.spark.cgescape.UnsafeExternalProcessorBase.UnsafeBatchProcessor
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.Attribute
+import org.apache.spark.sql.catalyst.expressions.UnsafeRow
 import org.apache.spark.sql.catalyst.expressions.codegen.CodeGenerator
 import org.apache.spark.sql.catalyst.expressions.codegen.CodegenContext
 import org.apache.spark.sql.catalyst.expressions.codegen.ExprCode
@@ -10,9 +12,24 @@ import org.apache.spark.sql.execution.BlockingOperatorWithCodegen
 import org.apache.spark.sql.execution.CodegenSupport
 import org.apache.spark.sql.execution.SparkPlan
 
+/**
+ * This is a base trait to provide us with batch processing capability that is melded with codegen.
+ * This will enable very high performance processing as the methods are very simple.
+ */
+object UnsafeExternalProcessorBase {
+  /** This is a dead-simple interface that receives UnsafeRows. Once all the rows are submitted, the caller will call
+   * .execute() to request results.
+   * TODO - Memory safety not yet implemented - TODO
+   * */
+  trait UnsafeBatchProcessor {
+    def insertRow(unsafeRow: UnsafeRow): Unit
+    def execute(): Iterator[InternalRow]
+  }
+}
 trait UnsafeExternalProcessorBase { this: SparkPlan with BlockingOperatorWithCodegen =>
   def child: SparkPlan
-  override protected def doExecute(): RDD[InternalRow] = sys.error("This should not be called if in WSCG")
+  override protected def doExecute(): RDD[InternalRow] =
+    sys.error("This should not be called if in WSCG")
   override def inputRDDs(): Seq[RDD[InternalRow]] = Seq(child.execute())
   type ContainerType <: UnsafeBatchProcessor
   def containerClass: Class[ContainerType]
@@ -22,7 +39,7 @@ trait UnsafeExternalProcessorBase { this: SparkPlan with BlockingOperatorWithCod
     val outputRow = ctx.freshName("outputRow")
     val thisPlan = ctx.addReferenceObj("plan", this)
     containerVariable = ctx.addMutableState(
-      classOf[UnsafeExternalDuplicator].getName,
+      containerClass.getName,
       "batchProcessor",
       v => s"$v = $thisPlan.createContainer();",
       forceInline = true
@@ -67,6 +84,5 @@ trait UnsafeExternalProcessorBase { this: SparkPlan with BlockingOperatorWithCod
        $containerVariable.insertRow((UnsafeRow)${row.value});
      """.stripMargin
   }
-
 
 }
