--- aurora4spark-parent/aurora4spark-sql-plugin/src/test/scala/com/nec/spark/agile/UnsafeBatchProcessor.scala
@@ -1,8 +1,0 @@
-package com.nec.spark.agile
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.UnsafeRow
-
-trait UnsafeBatchProcessor {
-  def insertRow(unsafeRow: UnsafeRow): Unit
-  def execute(): Iterator[InternalRow]
-}
